requirements.txt
# DB
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Django
Django==5.1
django-dotenv==1.4.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
setuptools==72.1.0

# Auth
dj-rest-auth[with_social]
djangorestframework-simplejwt==5.3.1

# Swagger
drf-yasg==1.21.7